% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automatic_resource_detection.R
\name{sim_auto_scale}
\alias{sim_auto_scale}
\title{Choose a similarity calculation method based on simple resource estimations}
\usage{
sim_auto_scale(
  n_rows,
  resources = NA,
  row2mem_coeff = 7.62940039228619e-06,
  row_cap = NA,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{n_rows}{Integer. The expected number of rows in the matrix for which we are calculating row similarity. \code{n_rows^2} will be used as the potential number of cells generated by the similarity matrix.}

\item{resources}{An optional list with \code{n_cpu} and \code{block_memory}. Default is \code{NA} which calls \code{\link[simscaleR:estimate_local_resources]{estimate_local_resources}}}

\item{row2mem_coeff}{Float. The coefficient that converts between \code{nrows^2} and object memory size in Mb. See \code{vignette('estimating-local-resources.Rmd', package = 'simscaleR')}.}

\item{row_cap}{Integer. Maximum number of rows for this machine. See \code{vignette('estimating-local-resources.Rmd', package = 'simscaleR')}. Default is \code{NA} which means no limit is applied}

\item{verbose}{Logical. Should informative messages be printed along the way?}

\item{...}{Additional parameters passed to \code{\link[simscaleR:estimate_local_resources]{estimate_local_resources}} (when \code{resources = NA})}
}
\value{
A function to calculate similarity with \code{n_cpu} set according to the recommendations. You need to provide \code{X, metric} and \code{thresh} if needed (defaults to 0). For more details see \code{\link[simscaleR:sim_loopR]{sim_loopR}})
}
\description{
Estimate number of CPU and memory size for the current system and use these numbers to choose the appropriate calculation method (loops / matrix multiplication, see \code{\link[simscaleR:sim_loopR]{sim_loopR}})
}
\examples{
\dontrun{sim_matrix <- auto_sim_scale(nrow(X))(X, 'cosine', thresh = 0.8)}
}
