% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{sim_loopR}
\alias{sim_loopR}
\alias{sim_blocksR}
\title{Generate similarity matrix using different methods}
\usage{
sim_loopR(X, metric, thresh = 0, n_cpu = 1, cl = NA)

sim_blocksR(X, metric, row_blocks = 1, thresh = 0, n_cpu = 1, cl = NA)
}
\arguments{
\item{X}{A matrix, or an object that can be coerced into a matrix. Values depend on the type of metric used (numeric, 1/0 integers, etc.)}

\item{metric}{One of the supported metrics. Currently \code{hamming} or \code{cosine}. For loop based functions you can also specify a function that takes two rows from the raw X and returns a distance}

\item{thresh}{Float. Minimal similarity threshold to be returned. Values below are converted to 0 (to allow sparse representation). Default is 0.0}

\item{n_cpu}{Integer. Number of cores to use for the local cluster (using the \code{doParallel} and \code{parallel} backend). Default is 1 which results in a simple R loop. Negative numbers are interpreted as "all CPU expect".}

\item{cl}{A cluster object, pointing to the cluster to be used instead of a local cluster. This option overrides the \code{n_cpu} parameter.}

\item{row_blocks}{Integer. How to divide matrix rows into blocks for block-wise similarity calculation (using matrix multiplication). Default is 1 which means similarity is calculated in a single step.}

\item{col_block}{Integer. Not implemented yet}
}
\value{
An object of \code{symmetricMatrix} class. Default is a sparse matrix but this can be modified using the \code{sparse} parameter
}
\description{
A collection of functions that take a numeric / 0-1 integer matrix and calculate row-wise similarity.
}
\section{Functions}{
\itemize{
\item \code{sim_loopR}: Generate sparse similarity matrix using simple R loop in a single thread or a cluster

\item \code{sim_blocksR}: Generate sparse similarity matrix using block-wise matrix multiplication
}}

\keyword{internal}
