% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_shuffle.R
\name{sim_matrix_shuffle}
\alias{sim_matrix_shuffle}
\title{Re-order a similarity matrix}
\usage{
sim_matrix_shuffle(X, row_order)
}
\arguments{
\item{X}{Matrix}

\item{row_order}{Integer vector, a permutation of \code{1:nrow(m)}.}
}
\value{
A re-arranged similarity matrix (sparse & symmetric)
}
\description{
Re-order columns / rows of a similarity matrix based on a given permutation.
}
