% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_matrix.R
\name{sparse_block_matrix}
\alias{sparse_block_matrix}
\title{Sparse block-matrix builder}
\usage{
sparse_block_matrix(blocks, values = 1)
}
\arguments{
\item{blocks}{A vector of integers corresponding to block sizes, from top left to bottom right (along the diagonal)}

\item{values}{A vector of values for each block or a single value for all blocks. Default is 1}
}
\description{
build a sparse block matrix of 1's along the diagonal, based on a vector of block sizes
}
