% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_by_partition.R
\name{merge_by_partition}
\alias{merge_by_partition}
\title{Merge matrix rows by a partition}
\usage{
merge_by_partition(m, partition, agg = Matrix::colSums)
}
\arguments{
\item{m}{A matrix}

\item{partition}{A list of integer vectors. Each vector can contain all rows in the group or just first & last rows (min/max will be used to extract range)}

\item{agg}{Function. The function that will be called to merge the rows. Must return a matrix with the same number of columns and as many rows as needed. Default is `colSums`.}
}
\value{
A new matrix with less (or more?) rows, depending on the aggregation function.
}
\description{
Merge matrix rows according to a continuous partition of row indices
}
