% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automatic_scale.R
\name{sim_auto_scale}
\alias{sim_auto_scale}
\title{Choose a similarity calculation method based on simple resource estimations}
\usage{
sim_auto_scale(
  n_rows,
  resources = NA,
  reg_coeff = c(-4.76483, 0.93864),
  verbose = FALSE,
  ...
)
}
\arguments{
\item{resources}{An optional list with \code{n_cpu} and \code{block_memory}. Default is \code{NA} which calls \code{\link[simscaleR:estimate_local_resources]{estimate_local_resources}}}

\item{verbose}{Logical. Should informative messages be printed along the way?}

\item{...}{Additional parameters passed to \code{\link[simscaleR:estimate_local_resources]{estimate_local_resources}} (when \code{resources = NA})}

\item{memory_reg}{A numeric vector with 2 items. First place is the intercept and second place is the coefficient for log(10) of number of cells. See example.}
}
\value{
A function to calculate similarity with \code{n_cpu} set according to the recommendations. You need to provide \code{X, metric} and \code{thresh} if needed (defaults to 0). For more details see \code{\link[simscaleR:sim_loopR]{sim_loopR}})
}
\description{
Estimate number of CPU and memory size for the current system and use these numbers to choose the appropriate calculation method (loops / matrix multiplication, see \code{\link[simscaleR:sim_loopR]{sim_loopR}})
}
\examples{
# How to calculate the regression coefficients on your system
memory_capacity <- data.frame(cells_log10 = 2:7, size = NA)

for (i in 1:nrow(memory_capacity)) {
  gc(full = TRUE)
  #!! try calc, on error NA
  memory_capacity[i, 'size_Mb'] <- tryCatch(
    as.numeric(object.size(Matrix::Matrix(data = 1.1, nrow = (10^memory_capacity[i, 'cells_log10'])/2, ncol = 2))) / (1024^2),
    error = function() NA
  )
  
  if (is.na(memory_capacity[i, 'size_Mb'])) {break}
}

\dontrun{plot(log10(size_Mb) ~ cells_log10, data = memory_capacity)}
mem_lm <- lm(log10(size_Mb) ~ cells_log10, data = memory_capacity)
summary(mem_lm)

}
